オプションの調整など
