+++ examples/build-docker.sh
@@ -1,0 +1,6 @@
+#! /bin/sh
+
+prefix=tmp/
+
+docker build -t ${prefix}hello -f ./Dockerfile.hello .
+docker build -t ${prefix}listfiles -f ./Dockerfile.listfiles .
